week 8.2 middleware

+++ app/Http/Controllers/StoreController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use App\Models\ProductCategory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class StoreController extends Controller
 {
@@ -24,6 +25,9 @@ class StoreController extends Controller
 
     public function insert_product(Request $request)
     {
+        if(!Gate::allows('insert-product')) {
+            abort(403, 'Unauthorized action.');
+        }
         $request->validate([
             'name' => 'required|string|max:255',
             'details' => 'nullable|string',
@@ -73,6 +77,9 @@ class StoreController extends Controller
 
     public function update_product(Request $request, $product_id)
     {
+        if(!Gate::allows('edit-product')) {
+            abort(403, 'Unauthorized action.');
+        }
         $request->validate([
             'name' => 'required|string|max:255',
             'details' => 'nullable|string',
@@ -105,6 +112,9 @@ class StoreController extends Controller
 
     public function delete_product($product_id)
     {
+        if(!Gate::allows('delete-product')) {
+            abort(403, 'Unauthorized action.');
+        }
         $product = Product::findOrFail($product_id);
         $product->delete();
         return redirect()->route('store')->with('success', 'Product deleted successfully!');
